Address warnings.

+++ include/polatory/model.hpp
@@ -90,7 +90,7 @@ public:
 
   // Experimental function.
   void set_parameters(const std::vector<double>& params) {
-    if (params.size() != num_parameters())
+    if (static_cast<int>(params.size()) != num_parameters())
       throw common::invalid_argument("params.size() == " + std::to_string(num_parameters()));
 
     set_nugget(params[0]);
